Backend/requirements.txt
Django>=5.0
python-dotenv>=1.0
djangorestframework>=3.14
django-cors-headers>=4.3
PyMySQL>=1.1
tensorflow>=2.15.0
numpy>=1.24.0
pillow>=10.0.0
scikit-learn>=1.3.0
pytest>=7.4.0
